Modulo_1-TratamientoDato/EjerciciosExtra/007/subscriber/requirements.txt
Faker==16.0.0
kafka-python
mysql-connector-python
pandas==1.5.0
requests
